Optimize single tuple fetch from nbtree index
Before this patch, calling amgettuple on an nbtree index always reads all
tuples on the page before returning the first tuple. Consecutive calls do
not need to lock the buffer and this helps performance in general. However,
when only one tuple is required, there is a lot of extra processing,
which hurts performance for common queries with a LIMIT 1 clause.

This patch changes _bt_first, which gets called for the first index
tuple to be fetched, to read only the first tuple from a page.
Consecutive calls to _bt_next will read all tuples as before. This
greatly improves performance for statements that have LIMIT 1 and fetch
from an index.

Author: Floris van Nee
Reviewed-by: <empty so far>